FIx 1 bug regarding the publish channel is null when there was a previous subscription. Fix JsonSerializer to support inheritance serialization/deserialization in message object. Add method GetByConsumerId for CompositeSubscription Fix minor bugs + add more unit tests

+++ src/Burrow.Tests/Extras/PriorityQueueSetupDataTests/Constructor.cs
@@ -14,7 +14,7 @@ namespace Burrow.Tests.Extras.PriorityQueueSetupDataTests
             var data = new PriorityQueueSetupData(3);
 
             // Asert
-            Assert.AreEqual(3, data.MaxPriorityLevel);
+            Assert.AreEqual((uint)3, data.MaxPriorityLevel);
             Assert.IsTrue(data.Durable);
         }
     }
